🐛 QV09NA effect-in-doubt: split resolution hot paths
