+++ .agentplane/tasks/202607242158-QV09NA/pr/meta.json
@@ -2,7 +2,7 @@
   "base": "main",
   "branch": "task/202607242158-QV09NA/resolve-durable-runner-effects-in-doubt-without",
   "created_at": "2026-07-27T05:36:45.894Z",
-  "diffstat_sha256": "sha256:201b06e32de1d149fe93e3c93e6dea7d61d72c8093324f81571609a0e487a362",
+  "diffstat_sha256": "sha256:47d3262afc3921e5945aaec92f3c2a17749799fcd3c6d0912c90dfdc72c7af77",
   "last_verified_at": "2026-07-27T06:38:24.275Z",
   "last_verified_diffstat_sha256": "sha256:d4730a63b96774efce99db488d1f49df77bf1c817ef872e2933c194afd682ca0",
   "pr_number": 4639,
