+++ packages/agentplane/src/runner/usecases/task-run-active-claim.ts
@@ -14,8 +14,6 @@ import {
 } from "../stable-file.js";
 import { resolveSupervisorTaskRunnerPaths } from "../task-run-paths.js";
 import type { RunnerLifecycleStatus } from "../types.js";
-import type { RunnerEffectResolutionRef } from "@agentplaneorg/core/schemas";
-import { RunnerRunRepository } from "../run-repository.js";
 
 import {
   inspectTaskRunnerActiveClaimOwner,
@@ -88,7 +86,7 @@ export type TaskRunnerActiveClaimRecoveryPreparation =
   | { status: "ready"; token: TaskRunnerActiveClaimRecoveryToken }
   | Extract<TaskRunnerActiveClaimRecoveryResult, { status: "refused" }>;
 
-async function resolveClaimDirectory(
+export async function resolveClaimDirectory(
   opts: TaskRunnerActiveClaimPathOptions,
 ): Promise<TaskRunnerActiveClaimDirectory> {
   const paths = await resolveSupervisorTaskRunnerPaths(opts);
@@ -112,7 +110,7 @@ async function resolveClaimDirectory(
   };
 }
 
-async function readObservedClaim(
+export async function readObservedClaim(
   directory: TaskRunnerActiveClaimDirectory,
 ): Promise<ObservedTaskRunnerActiveClaim | null> {
   await directory.boundary.assertStable("before reading runner active claim");
@@ -174,7 +172,7 @@ function ownerIdentityMatches(
   );
 }
 
-function ownershipError(opts: {
+export function ownershipError(opts: {
   expected: ObservedTaskRunnerActiveClaim;
   observed?: TaskRunnerActiveClaim | null;
   archive_path?: string;
@@ -200,7 +198,7 @@ function ownershipError(opts: {
   });
 }
 
-async function retireObservedClaim(opts: {
+export async function retireObservedClaim(opts: {
   directory: TaskRunnerActiveClaimDirectory;
   expected: ObservedTaskRunnerActiveClaim;
   disposition: "released" | "stale";
@@ -523,132 +521,6 @@ export async function completeTaskRunnerActiveClaimRecovery(
   return archivePath
     ? { status: "recovered", claim: observed.claim, archive_path: archivePath }
     : { status: "absent" };
-}
-
-/**
- * The generic recovery path must continue to reject effect_in_doubt.  This is
- * the only retirement path that is allowed to cross that boundary, and only
- * after the run state durably contains the exact operator resolution supplied
- * by the caller.  It deliberately has no adapter dependency.
- */
-export async function retireTaskRunnerActiveClaimAfterEffectResolution(opts: {
-  git_root: string;
-  workflow_dir: string;
-  task_id: string;
-  run_id: string;
-  expected_generation: string;
-  resolution: RunnerEffectResolutionRef;
-  ensure_resolution_attached: () => Promise<boolean>;
-}): Promise<"absent" | "retired"> {
-  const acquisition = await acquireTaskRunnerActiveClaimRecoveryLease({
-    git_root: opts.git_root,
-    workflow_dir: opts.workflow_dir,
-    task_id: opts.task_id,
-    target_generation: opts.expected_generation,
-  });
-  if (acquisition.status === "busy") {
-    throw new CliError({
-      exitCode: 8,
-      code: "E_RUNTIME",
-      message: "Runner effect resolution is already retiring this active claim.",
-      context: {
-        reason: "runner_effect_resolution_retirement_busy",
-        task_id: opts.task_id,
-        run_id: opts.run_id,
-        generation: opts.expected_generation,
-      },
-    });
-  }
-  let succeeded = false;
-  try {
-    const directory = await resolveClaimDirectory(opts);
-    const observed = await readObservedClaim(directory);
-    if (!observed) {
-      succeeded = true;
-      return "absent";
-    }
-    if (
-      observed.claim.run_id !== opts.run_id ||
-      observed.claim.generation !== opts.expected_generation
-    ) {
-      throw ownershipError({
-        expected: observed,
-        observed: observed.claim,
-        detail: "the active claim no longer matches the resolved generation",
-      });
-    }
-    const ownerStatus = await inspectTaskRunnerActiveClaimOwner(observed.claim);
-    if (ownerStatus !== "stale") {
-      throw new CliError({
-        exitCode: 8,
-        code: "E_RUNTIME",
-        message: "Runner effect resolution refuses to retire an active or unverified claim owner.",
-        context: {
-          reason:
-            ownerStatus === "active"
-              ? "runner_effect_resolution_owner_active"
-              : "runner_effect_resolution_owner_unverified",
-          task_id: opts.task_id,
-          run_id: opts.run_id,
-          generation: opts.expected_generation,
-        },
-      });
-    }
-    const repository = await RunnerRunRepository.openExistingTaskRun({
-      git_root: opts.git_root,
-      workflow_dir: opts.workflow_dir,
-      task_id: opts.task_id,
-      run_id: opts.run_id,
-      storage: "supervisor",
-    });
-    let record = await repository.readRequiredRecord({
-      task_id: opts.task_id,
-      run_id: opts.run_id,
-    });
-    if (!record.state.effect_resolution) {
-      await opts.ensure_resolution_attached();
-      record = await repository.readRequiredRecord({
-        task_id: opts.task_id,
-        run_id: opts.run_id,
-      });
-    }
-    const attached = record.state.effect_resolution;
-    if (
-      attached?.digest !== opts.resolution.digest ||
-      attached.active_claim_generation !== opts.expected_generation ||
-      attached.operation_key !== opts.resolution.operation_key ||
-      attached.operation_digest !== opts.resolution.operation_digest
-    ) {
-      throw new CliError({
-        exitCode: 8,
-        code: "E_RUNTIME",
-        message: "Runner effect resolution is not durably attached to the exact active claim.",
-        context: {
-          reason: "runner_effect_resolution_state_attachment_missing",
-          task_id: opts.task_id,
-          run_id: opts.run_id,
-          generation: opts.expected_generation,
-        },
-      });
-    }
-    await beginTaskRunnerActiveClaimRetirement({
-      lease: acquisition.lease,
-      task_dir: directory.task_dir,
-      target_generation: opts.expected_generation,
-    });
-    const archived = await retireObservedClaim({
-      directory,
-      expected: observed,
-      disposition: "stale",
-    });
-    succeeded = true;
-    return archived ? "retired" : "absent";
-  } finally {
-    await releaseTaskRunnerActiveClaimRecoveryLease({
-      lease: acquisition.lease,
-      succeeded,
-    });
-  }
 }
 
 export async function recoverTaskRunnerActiveClaim(
